#vms-1188. Backup storage. Playback ut.
<==== Mercurial data ====>
Branch: x_vms_1188_backup_storage_playback_ut

+++ mediaserver_core/src/plugins/resource/server_archive/dualquality_helper.h
@@ -38,7 +38,7 @@ private:
         DeviceFileCatalog::FindMethod       findMethod,
         bool                                preciseFind,
         SearchStack                         &searchStack,
-        int                                 previousDistance,
+        qint64                              previousDistance,
         int                                 findEps
     );
